% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhlthemes.R
\docType{package}
\name{nhlthemes}
\alias{nhlthemes}
\alias{nhlthemes-package}
\title{nhlthemes: R Studio addin for setting editor theme to NHL team colors}
\description{
nhlthemes: R Studio addin for setting editor theme to NHL team colors
}
